backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
sqlalchemy==2.0.38
pymysql==1.1.1
python-dotenv==1.0.1
scikit-learn==1.6.1
textblob==0.18.0.post0
vaderSentiment==3.3.2
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
